+++ push_pypi.sh
@@ -1,0 +1,4 @@
+#!/bin/bash -e
+rm -rf dist
+python setup.py sdist bdist_wheel
+twine upload --repository-url https://upload.pypi.org/legacy/ ./dist/*